requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
slack-bolt==1.18.1
slack-sdk==3.26.1
jira==3.5.2
openai==1.3.7
pyautogen==0.2.0
requests==2.31.0
python-dateutil==2.8.2
email-validator==2.1.0
gunicorn==21.2.0
Werkzeug==2.3.7 